requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
antlr4-python3-runtime
anyio
attrs
boto3
botocore
certifi
cffi
charset-normalizer
click
contourpy
cryptography
cycler
datasets
dill
effdet
elasticsearch==8.17.2
elastic-transport==8.17.1
fastapi
filelock
fonttools
frozenlist
fsspec
h11
huggingface-hub
idna
iniconfig
iopath
Jinja2
jmespath
joblib
kiwisolver
layoutparser
MarkupSafe
matplotlib
mpmath
multidict
multiprocess
mysql-connector-python
networkx
numpy
omegaconf
opencv-python
packaging
pandas
pdf2image
pdfminer.six
pdfplumber
pillow
pluggy
portalocker
propcache
pyarrow
pycocotools
pycparser
pydantic
pydantic_core
pyparsing
pypdfium2
pysbd
pytest
python-dateutil
python-dotenv
pytz
PyYAML
regex
requests
s3transfer
safetensors
scikit-learn
scipy
sentence-transformers
setuptools
six
sniffio
starlette
sympy
threadpoolctl
timm
tokenizers
torch
torchaudio
torchvision
tqdm
transformers
# typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
# vila @ file:///Users/jeong-yonghun/Desktop/capstone/CV_Search/src/CV2DB/vila
xxhash
yarl
httpx
prometheus_client
google-api-core
google-auth
google-cloud-vision
googleapis-common-protos
grpcio
grpcio-status